# banking-app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/bank-api/credit-wallet": {
            "post": {
                "description": "Credit Wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "credit"
                ],
                "summary": "Credit Wallet",
                "parameters": [
                    {
                        "description": "Credit wallet request",
                        "name": "payment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http_handler.CrDrRequestBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/http_handler.CrDrResponseBody"
                        }
                    }
                }
            }
        },
        "/bank-api/debit-wallet": {
            "post": {
                "description": "Debit Wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "debit"
                ],
                "summary": "Debit Wallet",
                "parameters": [
                    {
                        "description": "Debit wallet request",
                        "name": "payment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http_handler.CrDrRequestBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/http_handler.CrDrResponseBody"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "http_handler.CrDrRequestBody": {
            "type": "object",
            "required": [
                "account_id",
                "amount",
                "reference"
            ],
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "amount": {
                    "type": "number"
                },
                "reference": {
                    "type": "string"
                }
            }
        },
        "http_handler.CrDrResponseBody": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                }
            }
        }
    }
}